Client secret now comes standard
