+++ source/Server.OctopusID/Configuration/OctopusIDConfigureCommands.cs
@@ -26,11 +26,6 @@ namespace Octopus.Server.Extensibility.Authentication.OctopusID.Configuration
             {
                 yield return option;
             }
-            yield return new ConfigureCommandOption($"{ConfigurationSettingsName}ClientSecret=", "Tell Octopus the shared secret to use for Octopus ID authentication requests.", v =>
-            {
-                ConfigurationStore.Value.SetClientSecret(v.ToSensitiveString());
-                Log.Info($"{ConfigurationSettingsName} ClientSecret set");
-            }, hide: true);
         }
     }
 }
