+++ source/Server.OpenIDConnect.Common/Configuration/IOpenIDConnectConfiguration.cs
@@ -1,9 +1,12 @@
+using Octopus.Data.Model;
+
 namespace Octopus.Server.Extensibility.Authentication.OpenIDConnect.Common.Configuration
 {
     public interface IOpenIDConnectConfiguration
     {
         string? Issuer { get; set; }
         string? ClientId { get; set; }
+        SensitiveString? ClientSecret { get; set; }
         string? Scope { get; set; }
         string? NameClaimType { get; set; }
         bool AllowAutoUserCreation { get; set; }
