--- source/Server.OpenIDConnect.Common/Configuration/IOpenIDConnectConfigurationWithClientSecret.cs
@@ -1,9 +1,0 @@
-using Octopus.Data.Model;
-
-namespace Octopus.Server.Extensibility.Authentication.OpenIDConnect.Common.Configuration
-{
-    public interface IOpenIDConnectConfigurationWithClientSecret : IOpenIDConnectConfiguration
-    {
-        SensitiveString? ClientSecret { get; set; }       
-    }
-}
