+++ source/Server.OpenIDConnect.Common/Configuration/OpenIDConnectConfigurationResource.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel;
 using Octopus.Server.Extensibility.Extensions.Infrastructure.Configuration;
+using Octopus.Server.MessageContracts;
 using Octopus.Server.MessageContracts.Attributes;
 
 namespace Octopus.Server.Extensibility.Authentication.OpenIDConnect.Common.Configuration
@@ -14,6 +15,11 @@ namespace Octopus.Server.Extensibility.Authentication.OpenIDConnect.Common.Confi
         [Description("Octopus instances unique authentication id, as provided by your Octopus account")]
         [Writeable]
         public virtual string? ClientId { get; set; }
+
+        [DisplayName("Client Secret")]
+        [Description("Shared secret for validating the authentication tokens")]
+        [Writeable]
+        public virtual SensitiveValue? ClientSecret { get; set; }
 
         [DisplayName("Allow Auto User Creation")]
         [Description("Tell Octopus to automatically create a user account when a person signs in for the first time with this identity provider")]
