+++ source/Server.OpenIDConnect.Common/Configuration/OpenIDConnectConfigureCommands.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Octopus.Data.Model;
 using Octopus.Diagnostics;
 using Octopus.Server.Extensibility.Extensions.Infrastructure.Configuration;
 using Octopus.Server.Extensibility.HostServices.Web;
@@ -60,6 +61,10 @@ namespace Octopus.Server.Extensibility.Authentication.OpenIDConnect.Common.Confi
             {
                 ConfigurationStore.Value.SetClientId(v);
                 Log.Info($"{ConfigurationSettingsName} ClientId set to: {v}");
+            }, hide: hide);
+            yield return new ConfigureCommandOption($"{ConfigurationSettingsName}ClientSecret=", $"Follow our documentation to find the Client Secret for {ConfigurationSettingsName}.", v =>
+            {
+                ConfigurationStore.Value.SetClientSecret(v.ToSensitiveString());
             }, hide: hide);
         }
